Revert "Move and rename some classes"
This reverts commit 253ad3e6d63c55c6517e5a40f7b92318359b28e4.
