+++ src/main/kotlin/io/github/derechtepilz/infinity/gamemode/Extensions.kt
@@ -1,7 +1,6 @@
 package io.github.derechtepilz.infinity.gamemode
 
 import io.github.derechtepilz.infinity.Infinity
-import io.github.derechtepilz.infinity.gamemode.gameclass.GameClass
 import io.github.derechtepilz.infinity.gamemode.serializer.ExperienceSerializer
 import io.github.derechtepilz.infinity.gamemode.serializer.HealthHungerSerializer
 import io.github.derechtepilz.infinity.gamemode.serializer.InventorySerializer
