+++ src/main/kotlin/io/github/derechtepilz/infinity/gamemode/gamestage/GameStage.kt
@@ -1,0 +1,13 @@
+package io.github.derechtepilz.infinity.gamemode.gamestage
+
+enum class GameStage {
+
+	STAGE_01,
+	STAGE_1,
+	STAGE_2,
+	STAGE_3,
+	STAGE_4,
+	STAGE_5,
+	STAGE_6
+
+}
